Moved tests to unit folder
