+++ serverJS/__tests__/unit/DocumentValidation/DocumentValidator.test.js
@@ -1,5 +1,5 @@
 /* eslint-env jest */
-const { validateTextDocument } = require('../../src/DocumentValidation/DocumentValidator')
+const { validateTextDocument } = require('../../../src/DocumentValidation/DocumentValidator')
 
 describe('Tests for validation', () => {
   test('this does not do much yet', () => {
